Split weight related functions into a header
